#278 fix : 앨범 패키지 DTO 위치 이동 및 Error Message 누락 수정
