--- src/main/java/cord/eoeo/momentwo/album/adapter/dto/AlbumSubTitleEditRequestDto.java
@@ -1,13 +1,0 @@
-package cord.eoeo.momentwo.album.adapter.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AlbumSubTitleEditRequestDto {
-    private long albumId;
-    private String subTitle;
-}
